LS: Move all code actions logic into the `ide::code_actions` module
commit-id:67889a57
